My first commit on C preprocessing

+++ 0x00-hello_world/6-size.c
@@ -9,10 +9,10 @@
 
 int main(void)
 {
-printf("Size of a char: %zu bytes(s)", (unsigned long)sizeof(char));
-printf("Size of an int: %zu bytes(s)", (unsigned long)sizeof(int));
-printf("Size of a long int: %zu bytes(s)", (unsigned long)sizeof(long int));
-printf("Size of a long long int: %zu bytes(s)", (unsigned long)sizeof(long long int));
-printf("Size of a float: %zu bytes(s)", (unsigned long)sizeof(float));
+printf("Size of a char: %zu bytes(s)", sizeof(char));
+printf("Size of an int: %zu bytes(s)", sizeof(int));
+printf("Size of a long int: %zu bytes(s)", sizeof(long int));
+printf("Size of a long long int: %zu bytes(s)", sizeof(long long int));
+printf("Size of a float: %zu bytes(s)", sizeof(float));
 return (0);
 }
